Save `Interface` as u8

+++ src/space/common/mod.rs
@@ -22,9 +22,9 @@ impl Status {
 }
 
 #[derive(Copy, Clone)]
-struct Interface(u16);
+struct Interface(u8);
 impl Interface {
-    fn new(interface: u16) -> Self {
+    fn new(interface: u8) -> Self {
         Self(interface)
     }
 }
